chore: Move SelectItem to separate file
